requirements.txt
Flask
Pillow
opencv-python
youtube-transcript-api
sumy
pytesseract
tesseract-ocr
libtesseract-dev
onnxruntime

